# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/website/templates/{templateKey}/apply": {
      "post": {
        "operationId": "WebsiteController_applyTemplate",
        "parameters": [
          {
            "name": "templateKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
...
    "/website/pages/{id}": {
      "put": {
        "operationId": "WebsiteController_updatePage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/pages/{id}/sections": {
      "post": {
        "operationId": "WebsiteController_addSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/pages/{id}/revisions": {
      "get": {
        "operationId": "WebsiteController_getPageRevisions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/pages/{id}/draft": {
      "put": {
        "operationId": "WebsiteController_savePageDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/pages/{id}/publish": {
      "post": {
        "operationId": "WebsiteController_publishPage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/pages/{id}/rollback": {
      "post": {
        "operationId": "WebsiteController_rollbackPage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
...
    "/website/assets/{id}": {
      "delete": {
        "operationId": "WebsiteController_deleteAsset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
...
    "/website/form-submissions": {
      "get": {
        "operationId": "WebsiteController_getFormSubmissions",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formKey",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/form-submissions/{id}/status": {
      "put": {
        "operationId": "WebsiteController_updateFormSubmissionStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/analytics": {
      "get": {
        "operationId": "WebsiteController_getAnalytics",
        "parameters": [
          {
            "name": "range",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
...
    "/website/domains/{id}/check": {
      "post": {
        "operationId": "WebsiteController_checkDomain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/domains/{id}/retry": {
      "post": {
        "operationId": "WebsiteController_retryDomain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/domains/{id}/set-primary": {
      "post": {
        "operationId": "WebsiteController_setPrimaryDomain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/domains/{id}/routing": {
      "put": {
        "operationId": "WebsiteController_updateDomainRouting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/website/domains/{id}": {
      "delete": {
        "operationId": "WebsiteController_deleteDomain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Website"
        ]
      }
    },
    "/public/website/{domain}": {
      "get": {
        "operationId": "PublicWebsiteController_getPublicWebsite",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicWebsite"
        ]
      }
    },
    "/public/website/{domain}/sitemap.xml": {
      "get": {
        "operationId": "PublicWebsiteController_getPublicSitemap",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicWebsite"
        ]
      }
    },
    "/public/website/{domain}/robots.txt": {
      "get": {
        "operationId": "PublicWebsiteController_getPublicRobots",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicWebsite"
        ]
      }
    },
    "/public/website/preview/{token}": {
      "get": {
        "operationId": "PublicWebsiteController_getPreviewWebsite",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicWebsite"
        ]
      }
    },
    "/public/website/{domain}/forms/{formKey}/submit": {
      "post": {
        "operationId": "PublicWebsiteController_submitForm",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PublicWebsite"
        ]
      }
    },
    "/public/website/{domain}/analytics": {
      "post": {
        "operationId": "PublicWebsiteController_trackAnalytics",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PublicWebsite"
        ]
      }
    },
...
  },
  "info": {
    "title": "Noxera Plus API",
    "description": "Noxera Plus platform APIs for super-admin, admin, and public onboarding.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}